Created new component to share with Hired. Refactored with this component
